Implement top nav gradient layout - Added card and linear layout gradient layout - add darkmode nav functionality - add settings, wallet selector and dex button - clean up code and rebase PR

+++ ui/page/components/coinformat.go
@@ -36,7 +36,7 @@ func formatBalance(gtx layout.Context, l *load.Load, amount string, mainTextSize
 
 	mainText, subText := amount[:startIndex], amount[startIndex:]
 
-	subTextSize := unit.Sp(int(mainTextSize) * int(scale))
+	subTextSize := unit.Sp(float32(mainTextSize) * scale)
 
 	return layout.Flex{Axis: layout.Horizontal, Alignment: layout.Baseline}.Layout(gtx,
 		layout.Rigid(func(gtx C) D {
